chore: add .opencommitignore file to exclude go.sum from version control
The .opencommitignore file is added to the project to exclude the go.sum
file from being tracked by version control. This is done to prevent
unnecessary changes in the go.sum file from being committed.
In the go.mod file, several dependencies are updated to their latest
versions. This ensures that the project is using the most up-to-date
versions of the dependencies, which may include bug fixes, performance
improvements, or new features.