chore(ci.yml): update runs-on to use ubuntu-latest for better compatibility
The changes in the ci.yml file are made to improve compatibility with
the latest versions of the dependencies and tools used in the CI
workflow. This ensures that the CI workflow runs smoothly and is
up-to-date with the latest features and bug fixes. Additionally, the
tags in the docker build and push step are updated to include the branch
name and short sha for better identification of the built and pushed
images.