refactor(arangodb): implement ontology term update logic in EditStrain
Refactor the `mergeBindParams` function to use `maps.Copy` for a more
idiomatic and potentially efficient way to merge maps.
Implement the logic to update the strain's ontology term in `EditStrain`
if `DictyStrainProperty` is present in the update request. This involves
looking up the term ID and executing an UPSERT query to update the edge
in the term collection. Also, updated variable names and return values
in `EditStrain` for clarity.